feat: Disable Proposals component on Home page
The Proposals component on the Home page has been temporarily disabled. This change is necessary to address a bug or improve performance. Once the issue is resolved or the performance improvements are implemented, the Proposals component can be re-enabled.

This commit disables the Proposals component by commenting out the relevant code in the Home.tsx file.

Ref: Proposals